requirements_win.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Deprecated==1.2.13
dnspython==2.1.0
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
pyparsing==3.0.6
redis==4.1.0
six==1.16.0
Werkzeug==2.0.2
wrapt==1.13.3
pywin32==303
